requirements.txt
streamlit
torch
torchvision
sentence-transformers
python-dotenv
langchain-groq
chromadb
watchdog